Revert "i dont understand"
This reverts commit 8ad76bd4c4306c2974fb599ff118eb606859e1e5.

+++ CryptoCurrency/app/src/main/java/com/example/cryptocurrency/models/Coin.kt
@@ -14,5 +14,5 @@ class Coin(
     @SerializedName("image")
     val image:String,
     @SerializedName("total_volume")
-    val total_volume:Double
+    val total_volume:Long
 )
